# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cleaning company app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppUserRole = (typeof AppUserRole)[keyof typeof AppUserRole];

export const AppUserRole = {
  boss: "boss",
  employee: "employee",
} as const;

export interface AppUser {
  clerkId: string;
  username: string;
  email: string;
  role: AppUserRole;
  firstName?: string | null;
  lastName?: string | null;
  companyId?: number | null;
}

export interface Company {
  id: number;
  name: string;
  inviteCode: string;
  createdAt: string;
}

export interface CreateCompanyBody {
  name: string;
}

export interface JoinCompanyBody {
  inviteCode: string;
}

export type SetRoleBodyRole =
  (typeof SetRoleBodyRole)[keyof typeof SetRoleBodyRole];

export const SetRoleBodyRole = {
  boss: "boss",
  employee: "employee",
} as const;

export interface SetRoleBody {
  role: SetRoleBodyRole;
}

export type AssignmentStatus =
  (typeof AssignmentStatus)[keyof typeof AssignmentStatus];

export const AssignmentStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type AssignmentPriority =
  (typeof AssignmentPriority)[keyof typeof AssignmentPriority];

export const AssignmentPriority = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface Assignment {
  id: number;
  houseId: number;
  houseName: string;
  houseAddress: string;
  assignedToClerkId?: string | null;
  assignedToUsername?: string | null;
  date: string;
  timeSlot: string;
  notes?: string | null;
  guestCount?: number | null;
  status: AssignmentStatus;
  priority: AssignmentPriority;
  startedAt?: string | null;
  finishedAt?: string | null;
  createdAt: string;
}

export type CreateAssignmentBodyStatus =
  (typeof CreateAssignmentBodyStatus)[keyof typeof CreateAssignmentBodyStatus];

export const CreateAssignmentBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type CreateAssignmentBodyPriority =
  (typeof CreateAssignmentBodyPriority)[keyof typeof CreateAssignmentBodyPriority];

export const CreateAssignmentBodyPriority = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface CreateAssignmentBody {
  houseId: number;
  assignedToClerkId?: string | null;
  date?: string | null;
  timeSlot?: string | null;
  notes?: string | null;
  guestCount?: number | null;
  status: CreateAssignmentBodyStatus;
  priority: CreateAssignmentBodyPriority;
}

export type UpdateAssignmentBodyStatus =
  (typeof UpdateAssignmentBodyStatus)[keyof typeof UpdateAssignmentBodyStatus];

export const UpdateAssignmentBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type UpdateAssignmentBodyPriority =
  (typeof UpdateAssignmentBodyPriority)[keyof typeof UpdateAssignmentBodyPriority];

export const UpdateAssignmentBodyPriority = {
  low: "low",
  normal: "normal",
  high: "high",
} as const;

export interface UpdateAssignmentBody {
  houseId?: number | null;
  assignedToClerkId?: string | null;
  date?: string | null;
  timeSlot?: string | null;
  notes?: string | null;
  guestCount?: number | null;
  status?: UpdateAssignmentBodyStatus;
  priority?: UpdateAssignmentBodyPriority;
}

export type HouseStatus = (typeof HouseStatus)[keyof typeof HouseStatus];

export const HouseStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface House {
  id: number;
  name: string;
  mapLink?: string | null;
  ownerName: string;
  ownerPhone?: string | null;
  ownerEmail?: string | null;
  notes?: string | null;
  singleBeds?: number | null;
  doubleBeds?: number | null;
  babyBeds?: number | null;
  bathrooms?: number | null;
  jacuzzis?: number | null;
  saunas?: number | null;
  entryCode?: string | null;
  status: HouseStatus;
  createdAt: string;
}

export type CreateHouseBodyStatus =
  (typeof CreateHouseBodyStatus)[keyof typeof CreateHouseBodyStatus];

export const CreateHouseBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface CreateHouseBody {
  name: string;
  mapLink?: string | null;
  ownerName?: string | null;
  ownerPhone?: string | null;
  ownerEmail?: string | null;
  notes?: string | null;
  singleBeds?: number | null;
  doubleBeds?: number | null;
  babyBeds?: number | null;
  bathrooms?: number | null;
  jacuzzis?: number | null;
  saunas?: number | null;
  entryCode?: string | null;
  status: CreateHouseBodyStatus;
}

export type UpdateHouseBodyStatus =
  (typeof UpdateHouseBodyStatus)[keyof typeof UpdateHouseBodyStatus];

export const UpdateHouseBodyStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface UpdateHouseBody {
  name?: string;
  mapLink?: string | null;
  ownerName?: string;
  ownerPhone?: string | null;
  ownerEmail?: string | null;
  notes?: string | null;
  singleBeds?: number | null;
  doubleBeds?: number | null;
  babyBeds?: number | null;
  bathrooms?: number | null;
  jacuzzis?: number | null;
  saunas?: number | null;
  entryCode?: string | null;
  status?: UpdateHouseBodyStatus;
}

export interface PatchTimingBody {
  startedAt?: string | null;
  finishedAt?: string | null;
}

export interface UpdateHouseNotesBody {
  notes: string | null;
}

export interface HouseStats {
  total: number;
  active: number;
  inactive: number;
  weekly: number;
  biweekly: number;
  monthly: number;
  totalAssignmentsToday: number;
}