# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cleaning company app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company (boss only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
...
  /companies/join:
    post:
      operationId: joinCompany
      tags: [companies]
      summary: Join a company with invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinCompanyBody"
      responses:
        "200":
          description: Joined company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments (employees see their own, boss sees all)
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create a new assignment (boss only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get a single assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    put:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentBody"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assignments/{id}/timing:
    patch:
      operationId: patchAssignmentTiming
      tags: [assignments]
      summary: Manually adjust or clear the startedAt / finishedAt timestamps
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTimingBody"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}/start:
    post:
      operationId: startCleaning
      tags: [assignments]
      summary: Mark cleaning as started (records startedAt timestamp)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}/finish:
    post:
      operationId: finishCleaning
      tags: [assignments]
      summary: Mark cleaning as finished (records finishedAt timestamp)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        inviteCode:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - inviteCode
        - createdAt
    CreateCompanyBody:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    JoinCompanyBody:
      type: object
      properties:
        inviteCode:
          type: string
      required:
        - inviteCode
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        houseId:
          type: integer
        houseName:
          type: string
        houseAddress:
          type: string
        assignedToClerkId:
          type: string
          nullable: true
        assignedToUsername:
          type: string
          nullable: true
        date:
          type: string
        timeSlot:
          type: string
        notes:
          type: string
          nullable: true
        guestCount:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, normal, high]
        startedAt:
          type: string
          nullable: true
        finishedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - houseId
        - houseName
        - houseAddress
        - date
        - timeSlot
        - status
        - priority
        - createdAt
    CreateAssignmentBody:
      type: object
      properties:
        houseId:
          type: integer
        assignedToClerkId:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        timeSlot:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        guestCount:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, normal, high]
      required:
        - houseId
        - status
        - priority
    UpdateAssignmentBody:
      type: object
      properties:
        houseId:
          type: integer
          nullable: true
        assignedToClerkId:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        timeSlot:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        guestCount:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed]
        priority:
          type: string
          enum: [low, normal, high]
...
    PatchTimingBody:
      type: object
      properties:
        startedAt:
          type: string
          nullable: true
        finishedAt:
          type: string
          nullable: true
...